-Provided doc strings for camera package -Renamed a few files for logic and consistency. -moved digital camera shell scripts to a directory.
