+++ labvision/camera/scripts/nikon_config
